:lipstick: util 함수들 모듈화시키기
